pyproject.toml
[tool.poetry]
name = "fastapiAuth"
version = "0.1.0"
description = ""
authors = ["Oscar Giles <ogiles@turing.ac.uk>"]

[tool.poetry.dependencies]
python = "^3.7"
Authlib = "^0.15.3"
fastapi = {extras = ["all"], version = "^0.63.0"}
pydantic = "^1.8.1"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
httpx = "^0.17.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
isort = "^5.8.0"
pylint = "^2.7.2"
flake8 = "^3.9.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
